#108 updated integration tests

+++ storage-module/src/main/java/marowak/dev/entity/Character.java
@@ -29,11 +29,9 @@ public record Character(
         double y,
 
         @NotNull
-        int angle,
+        short angle,
 
         @Column(name = "is_online")
         boolean online
-
-        ) {
-
+) {
 }
